requirements.txt
# Core dependencies
click>=8.0.0
pyyaml>=6.0
aiofiles>=0.8.0
rich>=10.0.0

# Optional: Web monitoring (not yet implemented)
# fastapi>=0.68.0
# uvicorn>=0.15.0
# websockets>=10.0
# jinja2>=3.0

# Development dependencies
# pytest>=6.0
# pytest-asyncio>=0.18.0
# black>=22.0
# flake8>=4.0
# mypy>=0.910